requirements.txt
datasets==3.6.0
langchain-text-splitters==0.3.8
lightrag-hku==1.4.2
